move old fold code to separate file
