+++ networkExpansionPy/foldsOld.py
@@ -1,0 +1,81 @@
+import pandas as pd
+from pathlib import PurePath, Path
+from copy import deepcopy
+
+asset_path = PurePath(__file__).parent / "assets"
+
+# define a function that determins if a reaction rule (x) is feasible with foldSet
+def rule2rn(foldSet,x):
+    if x.issubset(foldSet):
+    #if foldSet.issubset(x)
+        return True
+    else:
+        return False
+
+# define a function that returns a list of reactions that are feasible with foldSet.  rules_sub contains all the fold set rules for all reactions
+def folds2rn(rules_sub,foldSet):
+    feasible = rules_sub['fold_sets'].apply(lambda x: rule2rn(foldSet,x))
+    y = rules_sub[feasible]
+    if len(y) > 0:
+        rns = rules_sub[feasible].rn.unique().tolist()
+    else:
+        rns = []
+    return rns
+
+# same as above but return only the set of rules that are feasible
+def folds2rules(rules_sub,foldSet):
+    feasible = rules_sub['fold_sets'].apply(lambda x: rule2rn(foldSet,x))
+    rule_df = rules_sub[feasible]
+    return rule_df
+
+
+# define a fold rules object, which simply contains fold sets that enable reactions
+class FoldRules:
+    
+    def __init__(self, path = PurePath("ecode","ecod2rn.ec3.07Feb2021.csv")):
+        # load the data
+        self.rules = None
+        self.rns = None 
+        self.folds = None
+
+    def copy(self):
+        return deepcopy(self)
+
+    def setRules(self,path = PurePath("ecode","ecod2rn.ec3.07Feb2021.csv")):
+        rules = pd.read_csv(asset_path / path, index_col=[0]).reset_index(drop=True)
+        self.rns = rules.rn.unique().tolist()
+        folds =  rules['rule'].apply(lambda x: set(x.split('_')))
+        folds = [item for sublist in folds for item in sublist]
+        self.folds = list(set(folds))
+        rules['fold_sets']= rules['rule'].apply(lambda x: set(x.split('_')))
+        self.rules = rules
+        
+    def folds2reactions(self,foldSet):
+        # make sure foldSet is of "set" type
+        if type(foldSet) != set:
+            if type(foldSet) != list:
+                foldSet = set([foldSet])
+            else:
+                foldSet = set(foldSet)
+                
+        rns = folds2rn(self.rules,foldSet)
+        return rns
+    
+    def removeFolds(self,folds_remove):
+        if type(folds_remove) != set:
+            if type(folds_remove) != list:
+                folds_remove = set([folds_remove])
+            else:
+                folds_remove = set(folds_remove)
+                
+        rulesRemoval = self.rules.fold_sets.apply(lambda x: folds_remove.issubset(x)) 
+        self.rules = self.rules[~rulesRemoval]
+        folds_remove_list = list(folds_remove)
+        self.folds = [x for x in self.folds if x not in folds_remove_list]
+
+#function that peforms fold expansion
+def fold_expansion(metabolism,foldRules,fold_set,cpds_set,rxns_seed):
+    rn_list = foldRules.folds2reactions(fold_set)
+    rn_list = metabolism.rxns2tuple(rn_list) + list(rxns_seed)
+    cx,rx = metabolism.expand(cpds_set,reaction_mask=rn_list)
+    return cx,rx,rn_list
